extracted common layout from login page and resonator feedback
